iplauction/views.py
from django.http import HttpResponse
# from django.views.decorators.csrf import csrf_exempt

from rest_framework.decorators import api_view
from rest_framework.renderers import JSONRenderer
from rest_framework.response import Response
from iplauction.serializers import PlayerSerializer, TeamSerializer

from .models import Player, Team


class JSONResponse(HttpResponse):
    """
    An HttpResponse that renders its content into JSON.
    """
    def __init__(self, data, **kwargs):
        content = JSONRenderer().render(data)
        kwargs['content_type'] = 'application/json'
        super(JSONResponse, self).__init__(content, **kwargs)


@api_view(['GET', ])
def team_list(request):
    team = Team.objects.all()
    serializer = TeamSerializer(team, many=True)
    return Response(serializer.data)


@api_view(['GET', ])
def player_list(request):
    player = Player.objects.all()
    serializer = PlayerSerializer(player, many=True)
    return Response(serializer.data)

