requirements.txt
Django==1.9.3
wheel==0.29.0
djangorestframework